LPS-74270 Rename ExceptionConverter to ExceptionMapper
The package change to exception.mapper can't be done until subrepo is merged
