+++ modules/apps/headless-apio/portal/portal-apio-impl/src/main/java/com/liferay/portal/apio/internal/architect/converter/NoSuchModelExceptionMapper.java
@@ -12,31 +12,31 @@
  * details.
  */
 
-package com.liferay.media.object.apio.internal.architect.converter;
+package com.liferay.portal.apio.internal.architect.converter;
 
 import com.liferay.apio.architect.converter.ExceptionMapper;
 import com.liferay.apio.architect.error.APIError;
-import com.liferay.document.library.kernel.exception.DuplicateFileEntryException;
+import com.liferay.portal.kernel.exception.NoSuchModelException;
 
 import javax.ws.rs.core.Response.Status;
 
 import org.osgi.service.component.annotations.Component;
 
 /**
- * Converts a {@code DuplicateFileEntryException} to its {@link APIError}
+ * Converts a {@code NoSuchModelException} to its {@link APIError}
  * representation.
  *
  * @author Alejandro Hernández
  */
 @Component(immediate = true)
-public class DuplicateFileEntryExceptionConverter
-	implements ExceptionMapper<DuplicateFileEntryException> {
+public class NoSuchModelExceptionMapper
+	implements ExceptionMapper<NoSuchModelException> {
 
 	@Override
-	public APIError map(DuplicateFileEntryException dfee) {
+	public APIError map(NoSuchModelException nsme) {
 		return new APIError(
-			dfee, "Duplicate entry", "bad-request",
-			Status.BAD_REQUEST.getStatusCode());
+			nsme, "Resource not found", "not-found",
+			Status.NOT_FOUND.getStatusCode());
 	}
 
 }
